move github events under source directory
