--- lib/daily_report_generator/github/events.rb
@@ -1,17 +1,0 @@
-# frozen_string_literal: true
-
-module DailyReportGenerator
-  module Github
-    # access github events
-    class Events
-      def initialize(client)
-        @client = client
-        @client.auto_paginate = true
-      end
-
-      def fetch
-        @client.user_events(@client.login)
-      end
-    end
-  end
-end
