+++ lib/daily_report_generator/source/events.rb
@@ -3,8 +3,8 @@
 require 'time'
 require 'octokit'
 require 'pry'
-require 'daily_report_generator/github/events'
-require 'daily_report_generator/github/event_adapter'
+require 'daily_report_generator/source/github/events'
+require 'daily_report_generator/source/github/event_adapter'
 require 'daily_report_generator/google_calendar/events'
 require 'daily_report_generator/google_calendar/event_adapter'
 
@@ -15,8 +15,8 @@ module DailyReportGenerator
       class << self
         def github
           octokit = Octokit::Client.new(netrc: true)
-          github_events = DailyReportGenerator::Github::Events.new(octokit).fetch
-          DailyReportGenerator::Github::EventAdapter.from(github_events)
+          github_events = DailyReportGenerator::Source::Github::Events.new(octokit).fetch
+          DailyReportGenerator::Source::Github::EventAdapter.from(github_events)
         end
 
         def google_calendar
