+++ lib/daily_report_generator/source/github/events.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+module DailyReportGenerator
+  module Source
+    module Github
+      # access github events
+      class Events
+        def initialize(client)
+          @client = client
+          @client.auto_paginate = true
+        end
+
+        def fetch
+          @client.user_events(@client.login)
+        end
+      end
+    end
+  end
+end
